fix: create new file for reset link and properly import firebase reset link method that has set config. Removed previous implementation of reset password link because example.js should be treated as a dummy file. Hence, why it is called example.js.
